разбивка шаблонов карточек
