+++ src/filters/create-template.js
@@ -1,9 +1,8 @@
-export default ({id, text, count = 0, active = false, additional = false}) => `
-  <a 
-    href="${id}" 
-    class="main-navigation__item 
-    ${active ? `main-navigation__item--active` : ``} 
-    ${additional ? `main-navigation__item--additional` : ``}">
-      ${text} 
-      ${count ? `<span class="main-navigation__item-count">${count}</span>` : ``}
-  </a>`;
+import {capitalizeFirstLetter} from '../utils';
+
+export default ({id, label = capitalizeFirstLetter(id), count = 0, isActive = false, additional = false}) => {
+  const className = `class="main-navigation__item 
+    ${isActive ? `main-navigation__item--active` : ``} 
+    ${additional ? `main-navigation__item--additional` : ``}`;
+  return `<a href="${id}" ${className}>${label} ${count ? `<span class="main-navigation__item-count">${count}</span>` : ``}</a>`;
+};
