+++ src/mok-data.js
@@ -1,0 +1,100 @@
+export const FilmsData = [
+  {
+    title: `The Assassination Of Jessie James By The Coward Robert Ford`,
+    rating: `9.8`,
+    info: {
+      year: `2018`,
+      duration: `1h 13m`,
+      genre: `Comedy`
+    },
+    src: `./images/posters/three-friends.jpg`,
+    description: `A priest with a haunted past and a novice on the threshold of her final vows are sent by the Vatican to investigate the death of a young nun in Romania and confront a malevolent force in the form of a demonic nun.`,
+    comments: `13 comments`,
+    controls: true
+  },
+  {
+    title: `Incredibles 2`,
+    rating: `9.8`,
+    info: {
+      year: `2018`,
+      duration: `1h 13m`,
+      genre: `Comedy`
+    },
+    src: `./images/posters/moonrise.jpg`,
+    description: `A priests Romania and confront a malevolent force in the form of a demonic nun.`,
+    comments: `13 comments`,
+    controls: true
+  },
+  {
+    box: `top`,
+    title: `Incredibles 2`,
+    rating: `9.8`,
+    info: {
+      year: `2018`,
+      duration: `1h 13m`,
+      genre: `Comedy`
+    },
+    src: `./images/posters/fuga-da-new-york.jpg`,
+    comments: `13 comments`
+  },
+  {
+    box: `top`,
+    title: `Incredibles 2`,
+    rating: `9.8`,
+    info: {
+      year: `2018`,
+      duration: `1h 13m`,
+      genre: `Comedy`
+    },
+    src: `./images/posters/blue-blazes.jpg`,
+    comments: `13 comments`
+  },
+  {
+    box: `commented`,
+    title: `Incredibles 2`,
+    rating: `9.8`,
+    info: {
+      year: `2018`,
+      duration: `1h 13m`,
+      genre: `Comedy`
+    },
+    src: `./images/posters/accused.jpg`,
+    comments: `13 comments`
+  },
+  {
+    box: `commented`,
+    title: `Incredibles 2`,
+    rating: `9.8`,
+    info: {
+      year: `2018`,
+      duration: `1h 13m`,
+      genre: `Comedy`
+    },
+    src: `./images/posters/blackmail.jpg`,
+    comments: `13 comments`
+  }
+];
+
+export const FiltersData = [
+  {
+    id: `all`,
+    label: `All movies`,
+    isActive: true
+  },
+  {
+    id: `watchlist`,
+    count: 13
+  },
+  {
+    id: `history`,
+    count: 4
+  },
+  {
+    id: `favorites`,
+    count: 8
+  },
+  {
+    id: `stats`,
+    additional: true
+  },
+];
